Fix medication package conversion bug
When retrieving a medication, its package is represented as a text in the
database. It was assumed that it would contain an amount and an unit. The first
would be an integer literal, and the second a text, both would be separated by a
space.

When working with the demo database, it appeared that several medications didn't
have a package complying with that format: the unit is sometimes not separated
from the integer by the space, or there is only a text (e.g. `EACH`).

This change fixes this bug as it follows:

* the integer litteral is extracted (thanks to a regexp),
* if there isn't one found, the amount 1 is returned.