Fix uncomplete Feedback creation (Claim side) (#144)
When creating a Feedback with the FHIR API, only the Feedback object is created.
However, on the Claim side, there are a few fields referencing the feedback,
its availability, and its status of the feedback claim (that seemed to be
SELECTED (i.e. int value `4`)). The present change improves the test to
check that this side is also correctly and adequately set, and the correct
update of the claim is also made.