Enforce `no` value in the DB for communication payload elements (#143)
A communication has a payload with 5 expected elements. When creating one, only
the `yes` value is stored. When reading, a value different of `yes` or its
absence is represented as a `no`. This leads to a record not enforcing the
`no` value in the database, which could be interpreted by other software
(e.g. another module such as the Claim module) incorrectly.

This change makes sur the `no` value is stored too.