Improve the test of communication API and check both cases (UUID, CODE) (#146)
* Improve the test of communication API and check both cases (UUID, CODE)

In a previous PR #136, we added the capacity to refer with CODE id rather than
UUID. A fixture was added to test this new case. However, as there is a 1-1
relation between Claim and Feedback, we couldn't test both fixtures, as a
Feedback would have been already linked to the Claim object used in the test.

We delete the created Feedback and it does the job.

---------

Co-authored-by: Patrick Delcroix <delcroip@gmail.com>